Make properties use floats for floating-point values

+++ src/core/properties/property.hpp
@@ -16,7 +16,7 @@ namespace tactile {
 template <typename T>
 concept IsPropertyType = std::same_as<T, std::string> ||
                          std::same_as<T, int> ||
-                         std::same_as<T, double> ||
+                         std::same_as<T, float> ||
                          std::same_as<T, bool> ||
                          std::same_as<T, cen::color> ||
                          std::same_as<T, std::filesystem::path> ||
@@ -29,7 +29,7 @@ class Property final
  public:
   using string_type = std::string;
   using integer_type = int;
-  using float_type = double;
+  using float_type = float;
   using color_type = cen::color;
   using file_type = std::filesystem::path;
